+++ bundles/org.openhab.ui/web/src/components/auth-mixin.js
@@ -44,11 +44,11 @@ export default {
             'code_verifier': codeVerifier
           })
 
-          this.$oh.setAccessToken(null)
+          this.$oh.api.setAccessToken(null)
           this.$oh.api.postPlain('/rest/auth/token?useCookie=true', payload, 'application/json', 'application/x-www-form-urlencoded').then((data) => {
             const resp = JSON.parse(data)
             localStorage.setItem('openhab.ui:refreshToken', resp.refresh_token)
-            this.$oh.setAccessToken(resp.access_token)
+            this.$oh.api.setAccessToken(resp.access_token)
             // schedule the next token refresh when 95% of this token's lifetime has elapsed, i.e. 3 minutes before a 1-hour token is due to expire
             setTimeout(this.refreshAccessToken, resp.expires_in * 950)
             this.$store.commit('setUser', { user: resp.user })
@@ -75,7 +75,7 @@ export default {
         this.$oh.setAccessToken(null)
         this.$oh.api.postPlain('/rest/auth/token', payload, 'application/json', 'application/x-www-form-urlencoded').then((data) => {
           const resp = JSON.parse(data)
-          this.$oh.setAccessToken(resp.access_token)
+          this.$oh.api.setAccessToken(resp.access_token)
           // schedule the next token refresh when 95% of this token's lifetime has elapsed, i.e. 3 minutes before a 1-hour token is due to expire
           setTimeout(this.refreshAccessToken, resp.expires_in * 950)
           // also make sure to check the token and renew it when the app becomes visible again
@@ -104,12 +104,12 @@ export default {
         localStorage.removeItem('openhab.ui:refreshToken')
         this.$oh.api.postPlain('/rest/auth/logout', payload, 'application/json', 'application/x-www-form-urlencoded').then((data) => {
           console.log('Logged out')
-          this.$oh.setAccessToken(null)
+          this.$oh.api.setAccessToken(null)
           this.$store.commit('setUser', { user: null })
           resolve()
         }).catch((err) => {
           console.log(err)
-          this.$oh.setAccessToken(null)
+          this.$oh.api.setAccessToken(null)
           this.$store.commit('setUser', { user: null })
           reject(err)
         })
