+++ bundles/org.openhab.ui/web/src/js/openhab/cordova/speech.js
@@ -1,0 +1,5 @@
+export default {
+  isRecognitionSupported () {
+    return false
+  }
+}
