+++ bundles/org.openhab.ui/web/src/js/openhab/index.js
@@ -1,0 +1,11 @@
+import api from './api'
+import sse from './sse'
+import media from './media'
+import speech from './speech'
+
+export default {
+  api,
+  sse,
+  media,
+  speech
+}
